Make BinaryFileBuffer public
